17-2 refactor : move tests -> test
