+++ backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/test/TestService.java
@@ -1,6 +1,7 @@
 package kr.ac.ajou.da.testhelper.test;
 
 import kr.ac.ajou.da.testhelper.account.Account;
+import kr.ac.ajou.da.testhelper.account.AccountMapper;
 import kr.ac.ajou.da.testhelper.test.definition.TestStatus;
 import kr.ac.ajou.da.testhelper.test.exception.TestNotFoundException;
 import kr.ac.ajou.da.testhelper.test.room.TestRoomService;
@@ -8,12 +9,18 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class TestService {
 
     private final TestRepository testRepository;
     private final TestRoomService testRoomService;
+    private final TestsMapper testsMapper;
+    private final AccountMapper accountMapper;
 
     @Transactional
     public Test getTest(Long testId) {
@@ -32,5 +39,21 @@ public class TestService {
             testRoomService.deleteRoomsForStudents(testId, updatedBy.getId());
         }
 
+    }
+
+    public List<HashMap<String, Object>> getTests(int accountId) throws SQLException {
+        String role = getAccountRole(accountId);
+
+        if(role.equals("PROFESSOR")) {
+            return testsMapper.getTestListOfProfessor(accountId);
+        } else if(role.equals("ASSISTANT")) {
+            return testsMapper.getTestListOfAssistant(accountId);
+        }
+
+        return null;
+    }
+
+    private String getAccountRole(int accountId) throws SQLException {
+        return accountMapper.getAccountRole(accountId);
     }
 }
